chore : 각 테스트에서 사용되는 Repository 분리
