+++ core/testing/src/commonMain/kotlin/com/whatever/caramel/core/testing/util/TestMessage.kt
@@ -1,6 +1,6 @@
 package com.whatever.caramel.core.testing.util
 
-data object TestConstants {
+data object TestMessage {
     const val NOT_USE_IN_TEST = "테스트에서 사용하지 않는 메소드입니다."
     const val REQUIRE_INIT_FOR_TEST = "해당 테스트를 위해서는 변수 초기화가 필요합니다."
 }
