+++ feature/splash/src/commonTest/kotlin/com/whatever/caramel/feature/splash/test/SplashViewModelTest.kt
@@ -2,19 +2,9 @@ package com.whatever.caramel.feature.splash.test
 
 import androidx.lifecycle.SavedStateHandle
 import app.cash.turbine.test
-import com.whatever.caramel.core.domain.entity.User
-import com.whatever.caramel.core.domain.exception.CaramelException
-import com.whatever.caramel.core.domain.exception.code.AuthErrorCode
-import com.whatever.caramel.core.domain.exception.code.NetworkErrorCode
-import com.whatever.caramel.core.domain.repository.AuthRepository
-import com.whatever.caramel.core.domain.repository.UserRepository
 import com.whatever.caramel.core.domain.usecase.user.RefreshUserSessionUseCase
-import com.whatever.caramel.core.domain.vo.auth.AuthToken
-import com.whatever.caramel.core.domain.vo.auth.SocialLoginType
-import com.whatever.caramel.core.domain.vo.auth.UserAuth
 import com.whatever.caramel.core.domain.vo.user.UserStatus
 import com.whatever.caramel.core.testing.domain.AuthTestFactory
-import com.whatever.caramel.core.testing.util.TestConstants
 import com.whatever.caramel.core.testing.util.assertEqualsWithMessage
 import com.whatever.caramel.feature.splash.SplashViewModel
 import com.whatever.caramel.feature.splash.mvi.SplashSideEffect
@@ -112,67 +102,5 @@ class SplashViewModelTest {
 
     companion object {
         const val REFRESH_TOKEN_ERROR_MSG = "refresh token error"
-    }
-}
-
-class FakeUserRepository : UserRepository {
-    var savedUserStatus: UserStatus? = null
-
-    override suspend fun getUserStatus(): UserStatus {
-        return savedUserStatus ?: throw CaramelException(
-            code = NetworkErrorCode.UNKNOWN,
-            message = TestConstants.REQUIRE_INIT_FOR_TEST
-        )
-    }
-
-    override suspend fun setUserStatus(status: UserStatus) {
-        savedUserStatus = status
-    }
-
-    override suspend fun createUserProfile(
-        nickname: String,
-        birthDay: String,
-        agreementServiceTerms: Boolean,
-        agreementPrivacyPolicy: Boolean
-    ): User {
-        throw UnsupportedOperationException(TestConstants.NOT_USE_IN_TEST)
-    }
-}
-
-class FakeAuthRepository : AuthRepository {
-    var saveAuthToken: AuthToken? = null
-    var refreshAuthTokenResponse: AuthToken? = null
-    var isRefreshFail = false
-
-    override suspend fun loginWithSocialPlatform(
-        idToken: String,
-        socialLoginType: SocialLoginType
-    ): UserAuth {
-        throw UnsupportedOperationException(TestConstants.NOT_USE_IN_TEST)
-    }
-
-    override suspend fun refreshAuthToken(oldToken: AuthToken): AuthToken {
-        if(isRefreshFail){
-            throw CaramelException(
-                code = AuthErrorCode.UNAUTHORIZED,
-                message = SplashViewModelTest.REFRESH_TOKEN_ERROR_MSG
-            )
-        }
-        return refreshAuthTokenResponse
-            ?: throw CaramelException(
-                code = NetworkErrorCode.UNKNOWN,
-                message = TestConstants.REQUIRE_INIT_FOR_TEST
-            )
-    }
-
-    override suspend fun saveTokens(authToken: AuthToken) {
-        saveAuthToken = authToken
-    }
-
-    override suspend fun getAuthToken(): AuthToken {
-        return saveAuthToken ?: throw CaramelException(
-            code = NetworkErrorCode.UNKNOWN,
-            message = TestConstants.REQUIRE_INIT_FOR_TEST
-        )
     }
 }
